feat: editor as its own tree, a game project template, and a web frontend
The editor moves out of src/editor + src/editor_app into a top-level editor/
directory behind TINA_BUILD_EDITOR, so an SDK consumer no longer configures
authoring code they cannot use. tools/ now runs before TinaGameSdkPackage,
because tina_configure_game_sdk_package() only installs targets that already
exist and would otherwise ship a package with no asset cooker.

Also lands:
- cmake/TinaGameProject.cmake and TinaNewProject.cmake plus templates/game-project,
  with the content/frontend split enforced at configure time rather than documented.
- An Emscripten html5 platform backend, samples/web, and a wasm32-emscripten preset.
- Scene-owned clear colour, Mesh3D emissive factor, and stable CSM cascade bounds
  (ADR 0042-0044), with BgfxClearColor split out and covered by tests.
- ContentRoot in core/io, gameplay test executables, and the 2d_terraria / 3d_voxel samples.

The template has not been verified end to end from an installed SDK prefix yet.

+++ tools/bench/UIBenchmarkWorkloads.cpp
@@ -1669,8 +1669,8 @@ void hashLogicalRect(DeterministicHash& hash, const UI::UILogicalRect& rect) noe
             accumulateCommitStatistics(statistics, report);
             report.motionSampledTracks += statistics.motion.lastSampledTrackCount;
             report.motionActiveTracks += statistics.motion.activeTrackCount;
-            report.motionTrackHighWater =
-                (std::max)(report.motionTrackHighWater, statistics.motion.trackHighWater);
+            report.motionTrackHighWater = (std::max)(report.motionTrackHighWater,
+                                                    static_cast<u64>(statistics.motion.trackHighWater));
             if (activeTarget == 0) {
                 ++report.motionZeroActiveIterations;
                 if (statistics.motion.lastSampledTrackCount != 0 ||
